+++ src/main/java/com/braintreegateway/PaymentMethodGateway.java
@@ -2,6 +2,7 @@ package com.braintreegateway;
 
 import com.braintreegateway.util.Http;
 import com.braintreegateway.util.NodeWrapper;
+import com.braintreegateway.util.PaymentMethodParser;
 import com.braintreegateway.exceptions.NotFoundException;
 
 public class PaymentMethodGateway {
@@ -62,32 +63,6 @@ public class PaymentMethodGateway {
     }
 
     public Result<? extends PaymentMethod> parseResponse(NodeWrapper response) {
-        if (response.getElementName() == "paypal-account") {
-            return new Result<PayPalAccount>(response, PayPalAccount.class);
-        } else if (response.getElementName() == "credit-card") {
-            return new Result<CreditCard>(response, CreditCard.class);
-        } else if (response.getElementName() == "europe-bank-account") {
-            return new Result<EuropeBankAccount>(response, EuropeBankAccount.class);
-        } else if (response.getElementName() == "apple-pay-card") {
-            return new Result<ApplePayCard>(response, ApplePayCard.class);
-        } else if (response.getElementName() == "android-pay-card") {
-            return new Result<AndroidPayCard>(response, AndroidPayCard.class);
-        } else if (response.getElementName() == "amex-express-checkout-card") {
-            return new Result<AmexExpressCheckoutCard>(response, AmexExpressCheckoutCard.class);
-        } else if (response.getElementName() == "coinbase-account") {
-            return new Result<CoinbaseAccount>(response, CoinbaseAccount.class);
-        } else if (response.getElementName() == "us-bank-account") {
-            return new Result<UsBankAccount>(response, UsBankAccount.class);
-        } else if (response.getElementName() == "venmo-account") {
-            return new Result<VenmoAccount>(response, VenmoAccount.class);
-        } else if (response.getElementName() == "visa-checkout-card") {
-            return new Result<VisaCheckoutCard>(response, VisaCheckoutCard.class);
-        } else if (response.getElementName() == "masterpass-card") {
-            return new Result<MasterpassCard>(response, MasterpassCard.class);
-        } else if (response.getElementName() == "samsung-pay-card") {
-            return new Result<SamsungPayCard>(response, SamsungPayCard.class);
-        } else {
-            return new Result<UnknownPaymentMethod>(response, UnknownPaymentMethod.class);
-        }
+        return PaymentMethodParser.parsePaymentMethod(response);
     }
 }