+++ src/main/java/com/braintreegateway/WebhookNotification.java
@@ -35,6 +35,7 @@ public class WebhookNotification {
         IDEAL_PAYMENT_COMPLETE("ideal_payment_complete"),
         IDEAL_PAYMENT_FAILED("ideal_payment_failed"),
         GRANTED_PAYMENT_INSTRUMENT_UPDATE("granted_payment_instrument_update"),
+        GRANTED_PAYMENT_INSTRUMENT_REVOKED("granted_payment_instrument_revoked"),
         LOCAL_PAYMENT_COMPLETED("local_payment_completed"),
         UNRECOGNIZED("unrecognized");
 
@@ -65,6 +66,7 @@ public class WebhookNotification {
     private ConnectedMerchantPayPalStatusChanged connectedMerchantPayPalStatusChanged;
     private IdealPayment idealPayment;
     private GrantedPaymentInstrumentUpdate grantedPaymentInstrumentUpdate;
+    private RevokedPaymentMethodMetadata revokedPaymentMethodMetadata;
     private LocalPaymentCompleted localPaymentCompleted;
     private String sourceMerchantId;
 
@@ -126,6 +128,10 @@ public class WebhookNotification {
 
         if (wrapperNode.findFirst("granted-payment-instrument-update") != null) {
             this.grantedPaymentInstrumentUpdate = new GrantedPaymentInstrumentUpdate(wrapperNode.findFirst("granted-payment-instrument-update"));
+        }
+
+        if (kind == WebhookNotification.Kind.GRANTED_PAYMENT_INSTRUMENT_REVOKED) {
+            this.revokedPaymentMethodMetadata = new RevokedPaymentMethodMetadata(wrapperNode);
         }
 
         if (wrapperNode.findFirst("local-payment") != null) {
@@ -199,6 +205,10 @@ public class WebhookNotification {
 
     public GrantedPaymentInstrumentUpdate getGrantedPaymentInstrumentUpdate() {
         return this.grantedPaymentInstrumentUpdate;
+    }
+
+    public RevokedPaymentMethodMetadata getRevokedPaymentMethodMetadata() {
+        return this.revokedPaymentMethodMetadata;
     }
 
     public LocalPaymentCompleted getLocalPaymentCompleted() {