+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.87.0
+* Add `GrantedPaymentMethodRevoked` webhook response objects
+* Add missing `getCardholderName` method to ApplePayCard class
+* Add missing nonce detail params for non-credit card nonces
+
 ## 2.86.0
 * Add `authorizationExpiresAt` to `Transaction`